requirements.txt
attr==0.3.2
Cython==3.0.8
bpython==0.24
brotli==1.1.0
ConfigParser==6.0.1
cryptography==41.0.3
cx_Oracle==8.3.0
dl==0.1.0
docutils==0.20.1
faiss_cpu==1.7.4
HTMLParser==0.0.2
importlib_metadata==6.8.0
ipdb==0.13.13
ipython==8.12.3
ipywidgets==8.1.2
jnius==1.1.0
keyring==24.3.0
lockfile==0.12.2
lxml==5.1.0
MySQL-python==1.2.5
nltk==3.8.1
numpy==1.26.4
ordereddict==1.1
Pillow==10.0.0
Pillow==10.2.0
protobuf==4.25.3
psycopg==3.1.18
psycopg2==2.9.9
pylibmc==1.6.3
pymemcache==4.0.0
pyOpenSSL==24.0.0
pytz==2023.3.post1
pywatchman==1.4.1
PyYAML==6.0.1
PyYAML==6.0.1
railroad==0.5.0
scikit_learn==1.4.1.post1
sentence_transformers==2.4.0
Sphinx==7.2.6
tblib==3.0.0
thread==0.1.3
toml==0.10.2
tornado==6.3.3
xmlrpclib==1.0.1
zipp==3.16.2
